Grab flights only if they are stored under the given confirmation code; ignore any other reservations
The page we are parsing may include not just flights with the given confirmation number, but also 'associated products' with different confirmation numbers. Prior to this fix, ReservationInfoParser pulled all of the flights, regardless of confirmation number. Because of this, the program sent an e-mail promising to check in to all associated flights, but was only able to check in to the flights with the given confirmation number.

This commit fixes the bug by ensuring that ReservationInfoParser only pulls flights for the given confirmation number.

At some point, we could add a feature, giving the user the option to check into all 'associated' flights. This should be easy. The 'associated products' (flights with differing confirmation numbers) are located inside the tag <h3 class="trip_associated_products">.