add first_name, last_name to reservation
Instead of trying to shoehorn these fields into the passenger table, it
makes sense to use the fields on the reservation as user input. A
`Passenger` corresponds to the response from SW that contains each
passenger on the reservation.

Therefore, first and last name exist in the reservations and passengers
table, but server different roles: user input and SW data, respectively.